# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "\u8ba2\u5355\u7cfb\u7edf",
        "version": "v0.0.1"
    },
    "basePath": "/lsx/public/index.php/v1",
    "paths": {
...
        "/lsx/public/index.php/v1/order/add": {
            "post": {
                "tags": [
                    "\u8ba2\u5355\u6a21\u5757"
                ],
                "summary": "\u4e0b\u5355\u63a5\u53e3",
                "description": "\u4e0b\u5355\u63a5\u53e3",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "foods",
                        "in": "formData",
                        "description": "\u83dc\u54c1\u6570\u7ec4",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "deskId",
                        "in": "formData",
                        "description": "\u9910\u684c\u7f16\u53f7",
                        "required": true,
                        "type": "number"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "\u8bf7\u6c42\u6210\u529f"
                    }
                }
            }
        },
        "/lsx/public/index.php/v1/order/list": {
            "get": {
                "tags": [
                    "\u8ba2\u5355\u6a21\u5757"
                ],
                "summary": "\u83b7\u53d6\u8ba2\u5355\u5217\u8868",
                "description": "\u83b7\u53d6\u8ba2\u5355\u5217\u8868,\u9700\u8981\u4f20\u9012\u9875\u6570\uff0c\u6bcf\u9875\u6761\u6570\uff0c\u8d77\u59cb\u65f6\u95f4\u548c\u7ed3\u675f\u65f6\u95f4,\u90fd\u662f\u53ef\u9009\u53c2\u6570\u6709\u9ed8\u8ba4\u503c",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "\u9875\u6570\uff0c\u4e0d\u4f20\u5219\u9ed8\u8ba4\u4e3a1",
                        "type": "string"
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "\u6bcf\u9875\u6761\u6570\uff0c\u4e0d\u4f20\u9ed8\u8ba4\u4e3a10",
                        "type": "string"
                    },
                    {
                        "name": "st",
                        "in": "query",
                        "description": "\u67e5\u8be2\u7684\u8d77\u59cb\u65f6\u95f4\uff0c\u4e0d\u4f20\u5219\u4e3a\u5f53\u524d\u65f6\u95f4-24\u5c0f\u65f6",
                        "type": "string"
                    },
                    {
                        "name": "et",
                        "in": "query",
                        "description": "\u67e5\u8be2\u7684\u7ed3\u675f\u65f6\u95f4\uff0c\u4e0d\u4f20\u5219\u4e3a\u5f53\u524d\u65f6\u95f4",
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "\u8bf7\u6c42\u6210\u529f"
                    }
                }
            }
        },
        "/lsx/public/index.php/v1/order/detail": {
            "get": {
                "tags": [
                    "\u8ba2\u5355\u6a21\u5757"
                ],
                "summary": "\u83b7\u53d6\u8ba2\u5355\u8be6\u60c5",
                "description": "\u83b7\u53d6\u8ba2\u5355\u8be6\u60c5",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "orderId",
                        "in": "query",
                        "description": "\u8ba2\u5355id",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "\u8bf7\u6c42\u6210\u529f"
                    }
                }
            }
        },
        "getChartInfo": {
            "get": {
                "tags": [
                    "\u8ba2\u5355\u6a21\u5757"
                ],
                "summary": "\u83b7\u53d6\u4e00\u5468\u8ba2\u5355\u6570\u91cf",
                "description": "\u83b7\u53d6\u4e00\u5468\u8ba2\u5355\u6570\u91cf\uff0c\u5df2\u7ed3\u8d26\uff0c\u5f85\u4ed8\u6b3e\uff0c\u5df2\u53d6\u6d88",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "time",
                        "in": "formData",
                        "description": "\u65f6\u95f4YYYY-mm-dd\u683c\u5f0f",
                        "required": false,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "\u8bf7\u6c42\u6210\u529f"
                    }
                }
            }
        },
        "/getMoneyInfo": {
            "get": {
                "tags": [
                    "\u8ba2\u5355\u6a21\u5757"
                ],
                "summary": "\u83b7\u53d6\u6bcf\u661f\u671f\u7684\u6210\u4ea4\u989d\u60c5\u51b5\u63a5\u53e3",
                "description": "\u83b7\u53d6\u6bcf\u661f\u671f\u7684\u6210\u4ea4\u989d\u60c5\u51b5\u63a5\u53e3",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "time",
                        "in": "formData",
                        "description": "\u65f6\u95f4YYYY-mm-dd\u683c\u5f0f",
                        "required": false,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "\u8bf7\u6c42\u6210\u529f"
                    }
                }
            }
        },
...
    },
...
}